+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.14.3
+
+* removed `GeoDB.Close`, as it is no longer required to close the file resource
+* fixed filtering by path for growth statistics
+* fixed concurrent access to GeoDB, the database file is now loaded into memory
+
 ## 1.14.2
 
 * optimized Postgres transaction handling